requirements.txt
numpy
flask
keras
Pillow
transformers
torch
tensorflow
gunicorn